requirements.txt
torch==1.8.0+cu102 torchvision==0.9.0+cu102 torchaudio===0.8.0 -f https://download.pytorch.org/whl/torch_stable.html
torch-optimizer
tensorboardX
scikit-learn
